finished login page and database integration
There is an issue with the welcome page where the data from the database being read is not being sent to the WelcomeActivity.java file (file which manages the activity welcome xml page), so it is not displaying the "Welcome [Y/N], your role is [role]" message. Log out button also needs programming.